largest number check fix

+++ 0x03-debugging/2-largest_number.c
@@ -8,9 +8,9 @@
  * Return: largest number
  */
 
-int largest_number(long a, long b, long c)
+int largest_number(int a, int b, int c)
 {
-	long largest;
+	int largest;
 
 		if (a > b && a > c)
 		{
